Refactor list page to use client-side task store and reactive rendering
